move helper in common
